fix(cfloat): move quad to uint64_t limbs, with a BlockType parity suite to keep it honest (#1588)
quad was the last large cfloat on 32-bit limbs; duble, xtndd and octo were
already uint64_t. On a cfloat<128,15> multiply-accumulate loop it is 1.47x
faster for an identical accumulated value.

Adds a BlockType parity suite. A cfloat's limb width is a storage choice --
how many blocks the encoding is cut into, how wide the carries are -- not part
of the format, so two cfloats with the same nbits and es must agree bit-for-bit
whatever their limb width. That makes the two configurations each other's
reference, and the suite needs no oracle. It compares uint8_t, uint16_t and
uint64_t against uint32_t over the encoding round-trip, add, sub, mul, div,
fma, the special encodings, 96-step ulp walks from both ends of the range,
subnormal arithmetic, and integer and double conversion. Operands come from an
explicit bit pattern applied to both types rather than two RNG streams assumed
to match, and include runs straddling every 8/16/32/64-bit edge. Formats are
chosen so the most significant limb is partially filled in several of them.
Verified to fail: a one-ulp perturbation of the candidate multiply fails 8541
cases and names the operands.

sqrt is excluded because both its branches are cfloat(std::sqrt((double)a)),
which touches no limb arithmetic. That is a defect in its own right, filed
separately.

Also fixes a pre-existing wrong-answer bug the work uncovered.
normalizeAddition's subnormal branch was missing the `fbits < 64` guard its
three sibling branches carry. fbits is unsigned, so for a wide cfloat the guard
is not loose but inverted: cfloat<128,15> has fbits = 112 and 64 - 112 wraps to
4294967248, so the 64-bit fast path ran on a 112-bit fraction, reading a
fraction_ull() that is 0 by design at that width and shifting it by up to 112.
minpos + minpos returned ZERO for every cfloat with fbits >= 64. UBSan in the
CI matrix named it; the parity assertion could not, because both limb widths
share fbits, took the same wrong branch and agreed with each other. A parity
suite finds defects that depend on the limb width and is blind to defects the
widths share.

large_types.cpp gains the standard regression-level default block. Without it
every REGRESSION_LEVEL_n was undefined outside CMake, the test body
preprocessed away, and it reported PASS having executed zero assertions.

BEHAVIOUR-CHANGE: cfloat subnormal addition is fixed for fbits >= 64. minpos +
minpos returned zero for xtndd, quad and octo and now returns the correct one
ulp; any subnormal sum in those formats was previously zero.

BEHAVIOUR-CHANGE: quad is re-aimed from cfloat<128,15,uint32_t> to
cfloat<128,15,uint64_t>. Values and results are unchanged, but quad is now a
different type, so an overload or template specialization keyed on the old
spelling no longer matches it.

Resolves #1587

+++ include/sw/universal/number/cfloat/cfloat.hpp
@@ -109,7 +109,7 @@ using fp64     = duble;
 using xtndd    = cfloat< 80, 11, uint64_t, true, false, false>;
 using fp80     = xtndd;
 // IEEE-754 quad (128bit) precision floating-point
-using quad     = cfloat<128, 15, uint32_t, true, false, false>;
+using quad     = cfloat<128, 15, uint64_t, true, false, false>;
 using fp128    = quad;
 // IEEE-754 octo (256bit) precision floating-point
 using octo     = cfloat<256, 19, uint64_t, true, false, false>;
